+++ lib/domain/error/error.dart
@@ -3,20 +3,6 @@
 // Date  : 2020/3/28
 // Time  : 11:25
 
-import 'package:verify/verify.dart';
-
-///
-/// 值对象验证错误
-/// 由错误的值对象的值在验证过程中返回的Error
-class ValidateError extends Error with ValidationError {
-  @override
-  final String errorDescription;
-  ValidateError(this.errorDescription);
-
-  @override
-  String toString() => 'ValidateError[$errorDescription]';
-}
-
 ///
 /// 包装Error
 abstract class MsgError<T> extends Error {
@@ -40,7 +26,7 @@ class NotAuthenticatedError extends Error {}
 ///
 /// 本地缓存出错
 class StorageError extends MsgError<String> {
-  StorageError({String msg}) : super('CacheFailure: $msg');
+  StorageError({String? msg}) : super('CacheFailure: $msg');
 }
 
 class TypeError extends MsgError<String> {