+++ lib/quick_start.dart
@@ -1,21 +1,22 @@
 library get_arch_quick_start;
 
+export 'package:get_arch_core/get_arch_core.dart';
+
+export 'qs_base.dart';
+
 export 'domain/error/error.dart';
 export 'domain/error/failures.dart';
-export 'domain/value_validators.dart';
-export 'package:verify/verify.dart' hide ValidationError;
 
 export 'interface/i_dialog.dart';
 export 'interface/i_network.dart';
 export 'interface/i_storage.dart';
 
 export 'infrastructure/ui/dialog_impl.dart';
-export 'infrastructure/ui/indicator.dart';
 export 'infrastructure/ui/views.dart';
+export 'infrastructure/ui/extension.dart';
 export 'infrastructure/network_impl.dart';
 export 'infrastructure/storage_impl.dart';
+export 'package:nil/nil.dart';
 
+export 'profile/get_arch_application_x.dart';
 export 'profile/quick_start_package.dart';
-
-export 'package:get_arch_core/get_arch_core.dart';
-export 'package:get_arch_core/profile/i_get_arch_package.dart';