+++ CHANGELOG.md
@@ -1,7 +1,35 @@
-## [0.7.0] -2020/7/8
-* feat(IHttp): 新增handleBytesRequest(),直接返回byte信息(用于获取图片bytes);
-* feat(IHttp): handleRequest()新增dynamic data参数;
-* feat(network_impl): 适配新的接口;
+## [1.1.0] -2020/8/14
+* notice: Only Injectable versions ">=1.0.0 <=1.0.2" are supported
+* fix(network_impl.dart): 修复网络请求没有发送"queryParameters"参数的bug!
+* dialog_impl:
+  feat: 当err()参数 failure为null时, 直接退出方法;
+  feat: 调整包结构, 允许分层导出 ( **可能不兼容旧版本的导包代码** , 尤其是单元测试);
+
+* add(extension.dart):
+  允许Failure, Future<Failure>快速调用 IDialog.err()方法创建对话框;
+  DialogX: errDialog();
+  FutureDialogX: asyncErrDialog();
+
+## [1.0.0] -2020/8/3  - Only Injectable versions ">=1.0.0 <=1.0.2" are supported!
+* 由于"injectable>1.0.2"版本有不兼容修改, 因此请不要在该版本引入">1.0.2"的injectable版本
+    如果出现bug, 请在 主项目的pubspec.yaml中添加 "injectable: '>=1.0.0 <=1.0.2'"
+* add(get_arch_application_x.dart):
+ 新增扩展方法runFlutter();
+ 新增扩展方法runMaterialApp();
+ 已兼容>3.0.0的BotToast;
+* refactor(failures.dart,dialog_impl.dart):
+  适配v1.0.0版本GetArchCore的UnknownFailure
+* feat(IHttp):
+ 新增handleBytesRequest(),直接返回byte信息(用于获取图片bytes);
+ handleRequest()新增dynamic data参数;
+* feat(network_impl):
+  适配新的接口;
+* refactor(dialog_impl/QuickDialog):
+  新增 FailureRoute,IDialog收到异常后,可以展示对应的DIalog;
+  新增 QuickAlter,快速构建美观的Dialog;
+  新增IDialog接口多个方法, 本类添加多个实现 selectTips(), err(), toast(), snack(), widget();
+* refactor(quick_start_package):
+  适配1.0.0以上的Injectable;
 
 ## [0.6.3] -2020/7/5
 * IStorage新增 Uint8List类型, int类型数据的读写, StorageImpl添加对应实现